rf(auction): reorganized the code base
