+++ src/auction/index.ts
@@ -1,17 +1,32 @@
-// Schemas
-export * from './schemas/nft.schema';
-export * from './schemas/auction.schema';
-export * from './schemas/auction-whitelist.schema';
-export * from './schemas/bid.schema';
-export * from './schemas/auction-history.schema';
-
-// Services
-export * from './auction.service';
-export * from './nft.service';
+// Module
+export { AuctionModule } from './auction.module';
 
 // Controllers
-export * from './auction.controller';
-export * from './nft.controller';
+export { AuctionController } from './controllers/auction.controller';
+export { NFTController } from './controllers/nft.controller';
+export { LifecycleController } from './controllers/lifecycle.controller';
+export { QueueController } from './controllers/queue.controller';
 
-// Module
-export * from './auction.module';
+// Services
+export { AuctionService } from './services/auction.service';
+export { NFTService } from './services/nft.service';
+export { AuctionLifecycleService } from './services/auction-lifecycle.service';
+export { AuctionNotificationService } from './services/auction-notification.service';
+export { BidQueueService } from './services/bid-queue.service';
+export { WebSocketAuthService } from './services/websocket-auth.service';
+
+// Gateways
+export { AuctionGateway } from './gateways/auction.gateway';
+
+// Schemas
+export { NFT, NFTSchema } from './schemas/nft.schema';
+export { Auction, AuctionSchema } from './schemas/auction.schema';
+export {
+  AuctionWhitelist,
+  AuctionWhitelistSchema,
+} from './schemas/auction-whitelist.schema';
+export { Bid, BidSchema } from './schemas/bid.schema';
+export {
+  AuctionHistory,
+  AuctionHistorySchema,
+} from './schemas/auction-history.schema';
