+++ src/auction/services/nft.service.ts
@@ -6,7 +6,7 @@ import {
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
-import { NFT, NFTStatus } from './schemas/nft.schema';
+import { NFT, NFTStatus } from '../schemas/nft.schema';
 
 /**
  * Service for managing NFTs in the auction system
